Introducing special ViewModel where values are set to be equal to the keys
that way switching to this special ViewModel is intended to make finding out which key is used in which place easier.

The special ViewModel is not included in the common options, but an own property in the "ICurrentViewModel". That way you would need to explicitly included it in your UI if wanted.